refactor: number 유효성 검사 위치 변경
